Rewrite SelectField
